Extract twitter feed to TwitterWidget
 This isoloation reduces the noise in the App.js file
